Move ResponseParsers to new file
